Revert "Feature/authentification"
