+++ .github/workflows/ci.yml
@@ -30,7 +30,6 @@ jobs:
         run: |
           echo "NEWS_API_KEY=${{ secrets.NEWS_API_KEY }}" >> .env
           echo "OPENAI_API_KEY=${{ secrets.OPENAI_API_KEY }}" >> .env
-          echo "NEXTAUTH_SECRET=${{ secrets.NEXTAUTH_SECRET }}" >> .env
 
       - name: Run tests
         run: npm test
