+++ app/[urlsegment]/page.tsx
@@ -8,12 +8,10 @@ import dynamic from 'next/dynamic';
 import { usePosts } from '@/context/NewsContext';
 import { useParams } from 'next/navigation';
 import MarkdownWrapper from '@/components/markdown-wrapper/MarkdownWrapper';
-import ProtectedRoute from '@/components/ProtectedRoute';
 
 const NewsAiContent = dynamic(() => import('./NewsAiContent'));
-const ChatBox = dynamic(() => import('@/components/chat-box/ChatBox'));
 
-const PostPage = () => {
+function PostPage() {
   const { posts } = usePosts();
   const { urlsegment } = useParams();
 
@@ -62,14 +60,13 @@ const PostPage = () => {
       <div className={styles.ArticleBody}>
         <MarkdownWrapper>{post.body.raw}</MarkdownWrapper>
       </div>
+
+      <br />
       <Suspense fallback={<div>Loading...</div>}>
-        <ProtectedRoute>
-          <NewsAiContent post={post} />
-          <ChatBox post={post} />
-        </ProtectedRoute>
+        <NewsAiContent post={post} />
       </Suspense>
     </div>
   );
-};
+}
 
 export default PostPage;
