--- components/SessionProviderWrapper.tsx
@@ -1,14 +1,0 @@
-'use client';
-
-import React from 'react';
-import { SessionProvider } from 'next-auth/react';
-
-const SessionProviderWrapper = ({
-  children,
-}: {
-  children: React.ReactNode;
-}) => {
-  return <SessionProvider>{children}</SessionProvider>;
-};
-
-export default SessionProviderWrapper;
