--- components/auth/AuthButton.module.css
@@ -1,81 +1,0 @@
-.authContainer {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.authText {
-  font-size: 14px;
-  color: #333;
-}
-
-.authIcon {
-  font-size: 20px;
-  color: #0070f3;
-  cursor: pointer;
-  transition: color 0.3s;
-}
-
-.authIcon:hover {
-  color: #005bb5;
-}
-
-.authButton {
-  background-color: #0070f3;
-  color: white;
-  border: none;
-  padding: 8px 16px;
-  border-radius: 4px;
-  cursor: pointer;
-  transition: background-color 0.3s;
-}
-
-.authButton:hover {
-  background-color: #005bb5;
-}
-
-.modalOverlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.modalContent {
-  background: white;
-  padding: 20px;
-  border-radius: 8px;
-  position: relative;
-  width: 300px;
-  max-width: 80%;
-}
-
-.closeButton {
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  font-size: 20px;
-  cursor: pointer;
-}
-
-.authForm {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.authInput {
-  padding: 8px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-}
-
-.authInput:focus {
-  outline: none;
-  border-color: #0070f3;
-}
