+++ components/navbar/Navbar.tsx
@@ -9,7 +9,6 @@ import { NavbarWrapper } from 'react-show-hide-sticky-navbar';
 import { useState } from 'react';
 
 import React from 'react';
-import AuthButton from '../auth/AuthButton';
 
 function Navbar() {
   const router = useRouter();
@@ -55,7 +54,6 @@ function Navbar() {
           <div className={styles.NavbarHeader}>
             <div className={styles.NavHeaderIconsLeft}>
               <ThemeButton />
-              <AuthButton />
             </div>
 
             <Link
