+++ hooks/useChatBox.ts
@@ -48,7 +48,6 @@ const useChatBox = ({ aiContent, initialQuestions }: UseChatBoxProps) => {
     isOpen,
     setIsOpen,
     suggestedQuestions,
-    setSuggestedQuestions,
     handleQuestionClick,
     toggleChatBox,
   };
