+++ package.json
@@ -20,7 +20,6 @@
     "date-fns": "^4.1.0",
     "express-rate-limit": "^7.5.0",
     "next": "^15.1.3",
-    "next-auth": "^4.24.11",
     "nextjs-toploader": "^3.7.15",
     "react": "^18",
     "react-dom": "^18",
