+++ styles/main.css
@@ -72,16 +72,7 @@ body {
   min-height: 100vh;
 }
 
-.Body {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-}
-
 .content {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
   padding-top: 3rem;
   width: 75%;
   height: 100%;
